+++ .commitmsg
@@ -1,1 +1,4 @@
-2402: AC.cpp (#1286) + word(en) - AC,35.71%,55.95%
+1351: AC.cpp (#1288) + (word(en))
+
+cpp - (printVersion)AC,100.00%,43.89%
+cpp - (少一次遍历)AC,100.00%,91.73%
