+++ README.md
@@ -545,6 +545,7 @@
 |1334.阈值距离内邻居最少的城市|中等|<a href="https://leetcode.cn/problems/find-the-city-with-the-smallest-number-of-neighbors-at-a-threshold-distance/" target="_blank">题目地址</a>|<a href="https://blog.letmefly.xyz/2023/11/14/LeetCode%201334.%E9%98%88%E5%80%BC%E8%B7%9D%E7%A6%BB%E5%86%85%E9%82%BB%E5%B1%85%E6%9C%80%E5%B0%91%E7%9A%84%E5%9F%8E%E5%B8%82/" target="_blank">题解地址</a>|<a href="https://letmefly.blog.csdn.net/article/details/134410277" target="_blank">CSDN题解</a>|<a href="https://leetcode.cn/problems/find-the-city-with-the-smallest-number-of-neighbors-at-a-threshold-distance/solutions/2527200/letmefly-1334yu-zhi-ju-chi-nei-lin-ju-zu-tz3g/" target="_blank">LeetCode题解</a>|
 |1338.数组大小减半|中等|<a href="https://leetcode.cn/problems/reduce-array-size-to-the-half/" target="_blank">题目地址</a>|<a href="https://blog.letmefly.xyz/2024/12/15/LeetCode%201338.%E6%95%B0%E7%BB%84%E5%A4%A7%E5%B0%8F%E5%87%8F%E5%8D%8A/" target="_blank">题解地址</a>|<a href="https://letmefly.blog.csdn.net/article/details/144487651" target="_blank">CSDN题解</a>|<a href="https://leetcode.cn/problems/reduce-array-size-to-the-half/solutions/3020818/letmefly-1338shu-zu-da-xiao-jian-ban-tan-4985/" target="_blank">LeetCode题解</a>|
 |1349.参加考试的最大学生数|困难|<a href="https://leetcode.cn/problems/maximum-students-taking-exam/" target="_blank">题目地址</a>|<a href="https://blog.letmefly.xyz/2023/12/26/LeetCode%201349.%E5%8F%82%E5%8A%A0%E8%80%83%E8%AF%95%E7%9A%84%E6%9C%80%E5%A4%A7%E5%AD%A6%E7%94%9F%E6%95%B0/" target="_blank">题解地址</a>|<a href="https://letmefly.blog.csdn.net/article/details/135217045" target="_blank">CSDN题解</a>|<a href="https://leetcode.cn/problems/maximum-students-taking-exam/solutions/2580213/letmefly-1349can-jia-kao-shi-de-zui-da-x-g7io/" target="_blank">LeetCode题解</a>|
+|1351.统计有序矩阵中的负数|简单|<a href="https://leetcode.cn/problems/count-negative-numbers-in-a-sorted-matrix/" target="_blank">题目地址</a>|<a href="https://blog.letmefly.xyz/2025/12/28/LeetCode%201351.%E7%BB%9F%E8%AE%A1%E6%9C%89%E5%BA%8F%E7%9F%A9%E9%98%B5%E4%B8%AD%E7%9A%84%E8%B4%9F%E6%95%B0/" target="_blank">题解地址</a>|<a href="https://letmefly.blog.csdn.net/article/details/156361831" target="_blank">CSDN题解</a>|<a href="https://leetcode.cn/problems/count-negative-numbers-in-a-sorted-matrix/solutions/3867784/letmefly-1351tong-ji-you-xu-ju-zhen-zhon-0dsb/" target="_blank">LeetCode题解</a>|
 |1366.通过投票对团队排名|中等|<a href="https://leetcode.cn/problems/rank-teams-by-votes/" target="_blank">题目地址</a>|<a href="https://blog.letmefly.xyz/2024/12/30/LeetCode%201366.%E9%80%9A%E8%BF%87%E6%8A%95%E7%A5%A8%E5%AF%B9%E5%9B%A2%E9%98%9F%E6%8E%92%E5%90%8D/" target="_blank">题解地址</a>|<a href="https://letmefly.blog.csdn.net/article/details/144814103" target="_blank">CSDN题解</a>|<a href="https://leetcode.cn/problems/rank-teams-by-votes/solutions/3033962/letmefly-1366tong-guo-tou-piao-dui-tuan-ssklk/" target="_blank">LeetCode题解</a>|
 |1367.二叉树中的链表|中等|<a href="https://leetcode.cn/problems/linked-list-in-binary-tree/" target="_blank">题目地址</a>|<a href="https://blog.letmefly.xyz/2024/12/30/LeetCode%201367.%E4%BA%8C%E5%8F%89%E6%A0%91%E4%B8%AD%E7%9A%84%E9%93%BE%E8%A1%A8/" target="_blank">题解地址</a>|<a href="https://letmefly.blog.csdn.net/article/details/144826418" target="_blank">CSDN题解</a>|<a href="https://leetcode.cn/problems/linked-list-in-binary-tree/solutions/3034359/letmefly-1367er-cha-shu-zhong-de-lian-bi-7zmc/" target="_blank">LeetCode题解</a>|
 |1373.二叉搜索子树的最大键值和|困难|<a href="https://leetcode.cn/problems/maximum-sum-bst-in-binary-tree/" target="_blank">题目地址</a>|<a href="https://blog.letmefly.xyz/2023/05/20/LeetCode%201373.%E4%BA%8C%E5%8F%89%E6%90%9C%E7%B4%A2%E5%AD%90%E6%A0%91%E7%9A%84%E6%9C%80%E5%A4%A7%E9%94%AE%E5%80%BC%E5%92%8C/" target="_blank">题解地址</a>|<a href="https://letmefly.blog.csdn.net/article/details/130779067" target="_blank">CSDN题解</a>|<a href="https://leetcode.cn/problems/maximum-sum-bst-in-binary-tree/solutions/2276880/letmefly-1373er-cha-sou-suo-zi-shu-de-zu-kldz/" target="_blank">LeetCode题解</a>|
