+++ Solutions/Other-English-LearningNotes-SomeWords.md
@@ -1703,6 +1703,7 @@ categories: [自用]
 |||
 |strap|n. 袋子，皮带<br/>捆扎|
 |novelette|n. 中篇小说|
+|||
 
 + 这个web要是能设计得可以闭眼(完全不睁眼)键盘控制背单词就好了。
 + 也许可以加个AI用最近词编故事功能(返回接口中支持标注所使用单词高亮?)
